requirements.txt
# Component A — LLM Serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1

# Component B — RAG Pipeline
chromadb==0.5.0
sentence-transformers==3.0.0
pypdf==4.2.0

# Component C — ASR
openai-whisper==20231117
torch==2.3.0
gtts==2.5.1

# Component D — PII Redaction
spacy==3.7.4

# Shared
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.2
click==8.1.7
typer==0.9.4